SSCS-7821: MYA – Select case – Display tab for active and non-active cases
