+++ app/server/controllers/login.ts
@@ -13,6 +13,7 @@ import { IdamService, TokenResponse, UserDetails } from '../services/idam';
 import { HearingService } from '../services/hearing';
 import { TrackYourApealService } from '../services/tyaService';
 import { Feature, isFeatureEnabled } from '../utils/featureEnabled';
+import { getHearingsByName } from '../utils/fieldValidation';
 
 const logger = Logger.getLogger('login.js');
 const idamUrlString: string = config.get('idam.url');
@@ -75,18 +76,6 @@ function redirectToIdam(idamPath: string, idamService: IdamService) {
     AppInsights.trackEvent('MYA_REDIRECT_IDAM_LOGIN');
     return res.redirect(idamUrl.href);
   };
-}
-
-function getHearingsByName(hearings) {
-  const hearingsByName = {};
-  hearings.forEach(hearing => {
-    const appellantName = hearing.appellant_name;
-    if (!hearingsByName[appellantName]) {
-      hearingsByName[appellantName] = [];
-    }
-    hearingsByName[appellantName].push(hearing);
-  });
-  return hearingsByName;
 }
 
 function getIdamCallback(
@@ -166,8 +155,8 @@ function getIdamCallback(
       } else {
         const hearingsByName = getHearingsByName(hearings);
         AppInsights.trackTrace(`[Cases count ${hearings.length}] - User logged in successfully as ${email}`);
-        req.session['hearingsByName'] = hearingsByName;
         if (isFeatureEnabled(Feature.MYA_PAGINATION_ENABLED, req.cookies)) {
+          req.session['hearings'] = hearings;
           return res.redirect(Paths.activeCases);
         } else {
           return res.render('select-case.html', { hearingsByName });
