+++ views/components/navigation-main-tabs.html
@@ -1,10 +1,10 @@
 {% macro navigationMainTabs(active) %}
 <div class="govuk-tabs" data-module="govuk-tabs">
-    <nav class="navigation-tabs"> 
+    <nav class="navigation-tabs">
     <ul class="govuk-tabs__list">
     {% for tab in mainTabs %}
-        <li class="govuk-tabs__list-item govuk-tabs__list-item--selected {{ "selected" if tab.id == active }}">
-            <a class="govuk-tabs__tab" style="text-decoration: underline;" id="tab-{{ tab.id }}" href="{{ tab.url }}">{{ tab.title | safe }}</a>
+        <li class="govuk-tabs__list-item{% if tab.id == active %} govuk-tabs__list-item--selected{% endif %}">
+            <a class="govuk-tabs__tab" role="tab" style="text-decoration: underline;" id="tab-{{ tab.id }}" href="{{ tab.url }}">{{ tab.title | safe }}</a>
         </li>
     {% endfor %}
     </ul>
