+++ views/dormant-tab.html
@@ -4,37 +4,39 @@
 
 {% block corContent %}
     <h1 class="govuk-heading-xl">{{ content[i18n.language].selectCase.heading | safe }}</h1>
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
         {{ navigationMainTabs("dormantTab") }}
-        <div class="task-list">
-            <h2 class="govuk-heading-m --margin-bottom-s">{{ content[i18n.language].dormantTab.header | safe }}</h2>
-        </div>
-        {% for name, hearings in hearingsByName %}
-            <h2 class="govuk-heading-m"><span id="appellant-name">{{ name | safe }}</span></h2>
-                <table class="govuk-table govuk-!-margin-bottom-9">
-                    <thead class="govuk-table__head">
-                        <tr class="govuk-table__row">
-                          <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.appealReference | safe }}</th>
-                          <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.benefit | safe }}</th>
-                          <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.dateSubmitted | safe }}</th>
-                          <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.mrnDate | safe }}</th>
-                          <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.action | safe }}</th>
-                        </tr>
-                    </thead>
-                    <tbody class="govuk-table__body">
-                        {% for hearing in hearings %}
-                            <tr class="govuk-table__row">
-                                <td class="govuk-table__cell">{{ hearing.case_reference | safe }}</td>
-                                <td class="govuk-table__cell">{{ content[i18n.language].benefitTypes[hearing.appeal_details.benefit_type.toLowerCase()].acronym | safe }}</td>
-                                <td class="govuk-table__cell">{{ hearing.appeal_details.submitted_date | dateFilter("locale", i18n.language) | dateFilter("DD MMMM YYYY") }}</td>
-                                <td class="govuk-table__cell">{{ hearing.appeal_details.mrn_date | dateFilter("locale", i18n.language) | dateFilter("DD MMMM YYYY") }}</td>
-                                <td class="govuk-table__cell"><a href="/sign-in?code=dummy&caseId={{ hearing.case_id }}">{{ content[i18n.language].selectCase.view | safe }}</a></td>
-                            </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
+        {% if dormantHearingByName|length %}
+            <div class="task-list">
+                <h2 class="govuk-heading-m --margin-bottom-s">{{ content[i18n.language].dormantTab.header | safe }}</h2>
             </div>
-        </div>
-    {% endfor %}
+            {% for name, hearings in dormantHearingsByName %}
+                <h2 class="govuk-heading-m"><span id="appellant-name">{{ name | safe }}</span></h2>
+                    <table class="govuk-table govuk-!-margin-bottom-9">
+                        <thead class="govuk-table__head">
+                            <tr class="govuk-table__row">
+                              <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.appealReference | safe }}</th>
+                              <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.benefit | safe }}</th>
+                              <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.dateSubmitted | safe }}</th>
+                              <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.mrnDate | safe }}</th>
+                              <th scope="col" class="govuk-table__header">{{ content[i18n.language].selectCase.action | safe }}</th>
+                            </tr>
+                        </thead>
+                        <tbody class="govuk-table__body">
+                            {% for hearing in hearings %}
+                                <tr class="govuk-table__row">
+                                    <td class="govuk-table__cell">{{ hearing.case_reference | safe }}</td>
+                                    <td class="govuk-table__cell">{{ content[i18n.language].benefitTypes[hearing.appeal_details.benefit_type.toLowerCase()].acronym | safe }}</td>
+                                    <td class="govuk-table__cell">{{ hearing.appeal_details.submitted_date | dateFilter("locale", i18n.language) | dateFilter("DD MMMM YYYY") }}</td>
+                                    <td class="govuk-table__cell">{{ hearing.appeal_details.mrn_date | dateFilter("locale", i18n.language) | dateFilter("DD MMMM YYYY") }}</td>
+                                    <td class="govuk-table__cell"><a href="/sign-in?code=dummy&caseId={{ hearing.case_id }}">{{ content[i18n.language].selectCase.view | safe }}</a></td>
+                                </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+            {% endfor %}
+        {% else %}
+            <div class="task-list">
+                <h2 class="govuk-heading-m --margin-bottom-s">{{ content[i18n.language].dormantTab.noDataMessage | safe }}</h2>
+            </div>
+        {% endif %}
 {% endblock %}
